test: migrate StudyAssetSummary to vitest
